requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
complimenter==1.0
flask-restful==0.3.9